refactor: quickview and F1 now bings quickdocs instead of ctrl-shift-e
