+++ src/base-config/keyboard.json
@@ -304,7 +304,7 @@
         "Ctrl-E"
     ],
     "navigate.toggleQuickDocs":  [
-        "Ctrl-Shift-E"
+        "F1"
     ],
     "navigate.previousMatch":  [
         {
@@ -325,6 +325,6 @@
         "Shift-F2"
     ],
     "help.support": [
-        "F1"
+        "Shift-F1"
     ]
 }
