+++ src/extensions/default/QuickView/QuickView.less
@@ -25,7 +25,7 @@
     background: @bc-panel-bg;
     position: absolute;
     z-index: 5000;
-    pointer-events: none;
+    user-select: text;
 
     padding: 8px;
     text-align: center;
